requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.0
requests==2.31.0
numpy>=1.21.0
pdfplumber>=0.10.0